chore: rename electrical_heating to opengeh_electrical_heating
