+++ source/electrical_heating/pyproject.toml
@@ -17,7 +17,7 @@ dependencies = [
 ]
 
 [project.scripts]
-execute = "electrical_heating.entry_point:execute"
+execute = "opengeh_electrical_heating.entry_point:execute"
 
 [dependency-groups]
 dev = [
@@ -36,12 +36,9 @@ build-backend = "hatchling.build"
 [tool.hatch.metadata]
 allow-direct-references = true
 
-[tool.hatch.build.targets.wheel]
-packages = ["src/electrical_heating"]
-
 [tool.pytest.ini_options]
 # -s ensures that print("<text>") is written to output. 
-addopts = "--cov-report=term --cov-report=xml:coverage.xml --cov=electrical_heating -ra -vv -s"
+addopts = "--cov-report=term --cov-report=xml:coverage.xml --cov=opengeh_electrical_heating -ra -vv -s"
 
 [tool.coverage.report]
 fail_under = 0
