+++ source/electrical_heating/src/opengeh_electrical_heating/domain/calculation_results.py
