+++ source/electrical_heating/src/opengeh_electrical_heating/domain/constants.py
