+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/electrical_heating_internal/repository.py
@@ -1,6 +1,6 @@
 ﻿from pyspark.sql import SparkSession, DataFrame
 
-from electrical_heating.infrastructure.electricity_market.database_definitions import (
+from opengeh_electrical_heating.infrastructure.electricity_market.database_definitions import (
     Database,
 )
