+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/electrical_heating_internal/schemas/calculations.py
