+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/electricity_market/__init__.py
