+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/electricity_market/data_values/metering_point_type.py
