+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/electricity_market/database_definitions.py
