+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/electricity_market/schemas/consumption_metering_point_periods_v1.py
