+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements_bronze/data_values/product.py
