+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements_bronze/database_definitions.py
