+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements_bronze/schemas/measurements_bronze_v1.py
