+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements_gold/column_names.py
