+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements_gold/data_values/metering_point_type.py
