+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements_gold/database_definitions.py
