+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/measurements_gold/schemas/time_series_points_v1.py
