+++ source/electrical_heating/src/opengeh_electrical_heating/infrastructure/spark_initializor.py
