+++ source/electrical_heating/tests/electrical_heating_tests/application/job_args/test_electrical_heating_job_args.py
@@ -2,11 +2,11 @@
 
 import pytest
 
-from electrical_heating.application.job_args.electrical_heating_job_args import (
+from opengeh_electrical_heating.application.job_args.electrical_heating_job_args import (
     parse_command_line_arguments,
     parse_job_arguments,
 )
-from electrical_heating.application.job_args.environment_variables import (
+from opengeh_electrical_heating.application.job_args.environment_variables import (
     EnvironmentVariable,
 )
